+++ src/quark.h
@@ -28,17 +28,18 @@
 #ifndef quark_h
 #define quark_h
 
-#include <stddef.h>
+#include <stdbool.h>
+#include <stdint.h>
 #include <stdio.h>
 
 typedef struct qrk_s qrk_t;
 
 qrk_t *qrk_new(void);
 void qrk_free(qrk_t *qrk);
-size_t qrk_count(const qrk_t *qrk);
+uint64_t qrk_count(const qrk_t *qrk);
 bool qrk_lock(qrk_t *qrk, bool lock);
-const char *qrk_id2str(const qrk_t *qrk, size_t id);
-size_t qrk_str2id(qrk_t *qrk, const char *key);
+const char *qrk_id2str(const qrk_t *qrk, uint64_t id);
+uint64_t qrk_str2id(qrk_t *qrk, const char *key);
 void qrk_load(qrk_t *qrk, FILE *file);
 void qrk_save(const qrk_t *qrk, FILE *file);
